pyproject.toml
[project]
name = "jenkins_helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.10.1",
    "python-jenkins>=1.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
jenkins_helper = "jenkins_helper.server:main"